VKT(Frontend): session expired modal for clerk UI [deploy]
