+++ frontend/packages/vkt/src/components/layouts/SessionStateHeader.tsx
@@ -8,10 +8,7 @@ import { Text } from 'shared/components';
 import { Color, Variant } from 'shared/enums';
 
 import { useCommonTranslation } from 'configs/i18n';
-import { useAppDispatch } from 'configs/redux';
 import { APIEndpoints } from 'enums/api';
-import { useInterval } from 'hooks/useInterval';
-import { loadPublicUser } from 'redux/reducers/publicUser';
 
 interface SessionStateHeaderProps {
   firstName: string;
@@ -22,16 +19,7 @@ export const SessionStateHeader: FC<SessionStateHeaderProps> = ({
   firstName,
   lastName,
 }) => {
-  const dispatch = useAppDispatch();
   const translateCommon = useCommonTranslation();
-
-  const heartBeat = () => {
-    if (!document.hidden) {
-      dispatch(loadPublicUser());
-    }
-  };
-
-  useInterval(heartBeat, 5000); // Every 5 seconds
 
   return (
     <div className="session-header columns gapped">
